Some experimental smoothing for the middle states.
Basic idea is to have the middle states "linger" for stuff like squinting and the like.  Right now it's a bit on the jittery side, so over multiple frames it'll crank up the smoothing.

Also made the open states configurable, so if you want resting bitch face all of the time, you can have resting bitch face all of the time.